updated to support the latest angular 2 changes
