+++ app/GlobalQuran/Api/Api.ts
@@ -1,10 +1,18 @@
-import {Injectable} from 'angular2/core';
-import {Http} from 'angular2/http'
+import {Injectable} from '@angular/core';
+import {Http} from '@angular/http';
+
+import {Observable} from 'rxjs/Rx';
+import 'rxjs/add/operator/map';
 
 @Injectable()
 export class Api {
+    static get parameters() {
+        return [[Http]];
+    }
 
-    constructor(private http: Http) {}
+    constructor(private http: Http) {
+
+    }
 
     getAll (surahNo:number, quranBy?:string)
     {
