+++ app/app.html
@@ -1,17 +1,1 @@
-<ion-menu id="leftMenu" [content]="content">
-
-  <ion-toolbar>
-    <ion-title>GlobalQuran</ion-title>
-  </ion-toolbar>
-
-  <ion-content>
-    <ion-list>
-      <button ion-item *ngFor="#p of pages" (click)="openPage(p)">
-        {{p.title}}
-      </button>
-    </ion-list>
-  </ion-content>
-
-</ion-menu>
-
-<ion-nav id="nav" [root]="rootPage" #content swipe-back-enabled="true"></ion-nav>
+<ion-nav light [root]="rootPage"></ion-nav>
