+++ app/pages/setting-transliteration/setting-transliteration.html
@@ -11,7 +11,7 @@
 <ion-content>
     <ion-list>
         <ion-item
-                *ngFor="#row of transliterationList"
+                *ngFor="let row of transliterationList"
                 [ngClass]="{'translationSelected': gq.isSelected(row.id)}"
                 (click)="gq.toggleSelect(row.id)">
             {{row.native_name}}
