+++ app/theme/app.md.scss
@@ -3,8 +3,8 @@
 
 // App Shared Variables
 // --------------------------------------------------
-// Shared Sass variables go in the app.varialbes.scss file
-@import 'app.variables';
+// Shared Sass variables go in the app.variables.scss file
+@import "app.variables";
 
 
 // App Material Design Variables
@@ -23,7 +23,7 @@
 // --------------------------------------------------
 // All Sass files that make up this app goes into the app.core.scss file.
 // For simpler CSS overrides, custom app CSS must come after Ionic's CSS.
-@import 'app.core';
+@import "app.core";
 
 
 // App Material Design Only Sass
