+++ app/theme/app.variables.scss
@@ -1,11 +1,10 @@
 // http://ionicframework.com/docs/v2/theming/
 
-
 // Ionic Shared Functions
 // --------------------------------------------------
 // Makes Ionic Sass functions available to your App
 
-@import 'globals.core';
+@import "globals.core";
 
 // App Shared Variables
 // --------------------------------------------------
