+++ app/theme/helpers/icons.scss
@@ -1,3 +1,5 @@
+$header-margin-svg: 0.6em;
+
 .fa-stack {
   position: relative;
   display: block;
@@ -55,7 +57,7 @@
 .svg-icon {
   z-index: 13;
   display: block;
-  margin: 1.5em auto;
+  margin: $header-margin-svg auto;
   position: relative;
   width: 100%;
   max-width: 2.5em;
@@ -64,7 +66,7 @@
 .logo-2 {
   z-index: 13;
   display: block;
-  margin: 1.5em auto;
+  margin: $header-margin-svg auto;
   position: relative;
   width: 100%;
   max-width: 7.5em;
