+++ ionic.project
@@ -1,0 +1,4 @@
+{
+  "name": "mobile-app",
+  "app_id": "5503dac7"
+}
