+++ package.json
@@ -1,25 +1,28 @@
 {
   "dependencies": {
-    "angular2": "2.0.0-beta.15",
-    "es6-promise": "3.0.2",
+    "@angular/common": "2.0.0-rc.3",
+    "@angular/compiler": "2.0.0-rc.3",
+    "@angular/core": "2.0.0-rc.3",
+    "@angular/platform-browser": "2.0.0-rc.3",
+    "@angular/platform-browser-dynamic": "2.0.0-rc.3",
+    "@angular/http": "2.0.0-rc.3",
     "es6-shim": "^0.35.0",
-    "ionic-angular": "2.0.0-beta.6",
-    "ionic-native": "^1.1.0",
-    "ionicons": "3.0.0-alpha.3",
-    "reflect-metadata": "0.1.2",
-    "ng2-vs-for": "1.0.11",
-    "rxjs": "5.0.0-beta.2",
-    "zone.js": "0.6.10"
+    "ionic-angular": "2.0.0-beta.10",
+    "ionic-native": "1.3.2",
+    "ionicons": "3.0.0",
+    "reflect-metadata": "^0.1.3",
+    "rxjs": "5.0.0-beta.6",
+    "zone.js": "^0.6.12"
   },
   "devDependencies": {
     "del": "2.2.0",
     "gulp": "3.9.1",
     "gulp-watch": "4.3.5",
-    "ionic-gulp-browserify-typescript": "^1.0.0",
+    "ionic-gulp-browserify-typescript": "2.0.0",
     "ionic-gulp-fonts-copy": "^1.0.0",
     "ionic-gulp-html-copy": "^1.0.0",
     "ionic-gulp-sass-build": "^1.0.0",
-    "ionic-gulp-scripts-copy": "^1.0.0",
+    "ionic-gulp-scripts-copy": "^2.0.0",
     "run-sequence": "1.1.5"
   },
   "cordovaPlugins": [
@@ -28,8 +31,7 @@
     "cordova-plugin-whitelist",
     "cordova-plugin-splashscreen",
     "cordova-plugin-statusbar",
-    "ionic-plugin-keyboard",
-    "cordova-plugin-nativeaudio"
+    "ionic-plugin-keyboard"
   ],
   "cordovaPlatforms": [
     "ios",
@@ -39,6 +41,6 @@
       "locator": "ios"
     }
   ],
-  "name": "GlobalQuran",
-  "description": "GlobalQuran APP"
+  "name": "Global Quran",
+  "description": "Global Quran"
 }
