--- typings/browser.d.ts
@@ -1,1 +1,0 @@
-/// <reference path="browser/ambient/es6-shim/es6-shim.d.ts" />
