+++ typings/index.d.ts
@@ -1,0 +1,1 @@
+/// <reference path="globals/es6-shim/index.d.ts" />
