--- typings/main.d.ts
@@ -1,1 +1,0 @@
-/// <reference path="main/ambient/es6-shim/es6-shim.d.ts" />
